Rename use-archive-subagent.pure.ts to archive-subagent.ts
